+++ app/src/main/java/com/antonio/samir/meteoritelandingsspots/features/list/MeteoriteList.kt
@@ -21,6 +21,7 @@ import androidx.paging.compose.itemKey
 import com.antonio.samir.meteoritelandingsspots.R
 import com.antonio.samir.meteoritelandingsspots.designsystem.ui.components.Loading
 import com.antonio.samir.meteoritelandingsspots.designsystem.ui.components.MessageError
+import com.antonio.samir.meteoritelandingsspots.designsystem.ui.components.MeteoriteItem
 import com.antonio.samir.meteoritelandingsspots.designsystem.ui.components.MeteoriteItemView
 import kotlinx.coroutines.flow.Flow
